DAACS/daacs_os/services/api/requirements.txt
# ===== DAACS OS API Dependencies =====

# Core
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.10.*
pydantic-settings==2.7.*

# Database
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
psycopg[binary]==3.2.*
aiosqlite==0.21.*
alembic==1.14.*

# Redis
redis[hiredis]==5.2.*

# Task Queue
celery[redis]==5.4.*

# LangGraph
langgraph==0.2.*
langgraph-checkpoint-postgres==2.0.*

# LLM Providers
httpx==0.28.*

# Config
pyyaml==6.0.*

# Utilities
python-dotenv==1.0.*
python-multipart==0.0.*
python-jose[cryptography]==3.*
email-validator==2.*
passlib[bcrypt]==1.*
bcrypt==4.0.*
cryptography==42.*
sentry-sdk[fastapi]==2.*
posthog==3.*

# Tests
pytest==8.*
pytest-asyncio==0.24.*
anyio[trio]==4.*
pytest-cov==6.*
